+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.schedule/src/com/mmxlabs/models/lng/schedule/impl/FuelUsageImpl.java
@@ -80,7 +80,7 @@ public class FuelUsageImpl extends EObjectImpl implements FuelUsage {
 	 * @generated
 	 */
 	@Override
-	public int getFuelCost() {
+	public double getFuelCost() {
 		int sum = 0;
 		for (final FuelQuantity fq : getFuels()) {
 			sum += fq.getCost();
