+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.schedule/src/com/mmxlabs/models/lng/schedule/impl/PurgeImpl.java
@@ -139,7 +139,7 @@ public class PurgeImpl extends EventImpl implements Purge {
 	 * @generated
 	 */
 	@Override
-	public int getFuelCost() {
+	public double getFuelCost() {
 		int sum = 0;
 		for (final FuelQuantity fq : getFuels()) {
 			sum += fq.getCost();
