+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.schedule/src/com/mmxlabs/models/lng/schedule/impl/SlotVisitImpl.java
@@ -452,7 +452,7 @@ public class SlotVisitImpl extends EventImpl implements SlotVisit {
 	 * @generated
 	 */
 	@Override
-	public int getFuelCost() {
+	public double getFuelCost() {
 		int sum = 0;
 		for (final FuelQuantity fq : getFuels()) {
 			sum += fq.getCost();
